msm8998: vdec: Add missing ifdefs for UBWC on DPB buffer decision
When the UBWC feature is available on a SoC, we define _UBWC_.
Vice-versa, when this is not defined, do not try to set any
UBWC mode as, otherwise, the Venus firmware may crash because
of unsupported modes, breaking the whole vidc functionality.

This happens on at least "legacy" SoCs, such as MSM8917,
MSM8939, MSM8952, MSM8956, MSM8976.

Change-Id: I76131db5272b97016679c5bc0bf6ae099167cd03